Adress most comments in review
Summary:
 - Broke ClassificationH2OModel into two subtypes: AnomalyDetectionClassificationH2OModel and SupervisedClassificationH2OModel
 - Updated commit message
 - among others
