+++ openml-datarobot/pom.xml
@@ -87,17 +87,6 @@
         </dependency>
     </dependencies>
 
-    <dependencyManagement>
-        <dependencies>
-            <dependency>
-                <groupId>com.feedzai</groupId>
-                <artifactId>openml-api</artifactId>
-                <version>${openml-api.version}</version>
-                <scope>provided</scope>
-            </dependency>
-        </dependencies>
-    </dependencyManagement>
-
     <build>
         <plugins>
             <plugin>
