+++ openml-h2o/src/main/java/com/feedzai/openml/h2o/algos/H2OIsolationForestUtils.java
@@ -1,10 +1,18 @@
 /*
- * The copyright of this file belongs to Feedzai. The file cannot be
- * reproduced in whole or in part, stored in a retrieval system,
- * transmitted in any form, or by any means electronic, mechanical,
- * photocopying, or otherwise, without the prior permission of the owner.
+ * Copyright 2019 Feedzai
  *
- * © 2019 Feedzai, Strictly Confidential
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
  */
 
 package com.feedzai.openml.h2o.algos;
@@ -38,7 +46,7 @@ public class H2OIsolationForestUtils extends AbstractUnsupervisedH2OParamUtils<I
             ParametersBuilderUtil.getParamSetters(IsolationForestParametersV3.class);
 
     @Override
-    IsolationForestParametersV3 parseSpecificParams(final IsolationForestParametersV3 h2oParams, final Map<String, String> params, final long randomSeed) {
+    protected IsolationForestParametersV3 parseSpecificParams(final IsolationForestParametersV3 h2oParams, final Map<String, String> params, final long randomSeed) {
         h2oParams.seed = randomSeed;
         params.forEach((paramName, value) -> cleanParam(value).ifPresent(paramValue ->
                 PARAMS_SETTER.setValueIn(h2oParams, paramName, paramValue))
@@ -47,7 +55,7 @@ public class H2OIsolationForestUtils extends AbstractUnsupervisedH2OParamUtils<I
     }
 
     @Override
-    IsolationForestParametersV3 getEmptyParams() {
+    protected IsolationForestParametersV3 getEmptyParams() {
         return new IsolationForestParametersV3().fillFromImpl();
     }
